+++ bika/lims/catalog/analysis_catalog.py
@@ -112,7 +112,6 @@ _columns_list = [
     "getInstrumentEntryOfResults",
     "getAllowedInstrumentUIDs",
     "getInstrumentUID",
-    "getResultsRange",
     "getSampleTypeUID",
     "getClientOrderNumber",
     "getDateReceived",