+++ bika/lims/upgrade/v01_03_003.py
@@ -238,6 +238,12 @@ METADATA_TO_REMOVE = [
     ("bika_setup_catalog", "cancellation_state"),
     ("bika_setup_catalog", "getName"),
     ("bika_setup_catalog", "getServiceUID"),
+
+    # Was only used in analyses listing, but it can lead to inconsistencies
+    # because there are some analyses (Duplicates) their result range depends
+    # on the result of an original analysis. Thus, better to remove the metadata
+    # and wake-up object than add additional reindexes, etc. everywhere
+    (CATALOG_ANALYSIS_LISTING, "getResultsRange")
 ]
 
 
@@ -653,4 +659,3 @@ def update_analyses_results_range(sample):
         if analysis_rr:
             analysis = api.get_object(analysis)
             analysis.setResultsRange(analysis_rr)
-            analysis.reindexObject()